refactor (addin): breaking filenaming changes
Trying to get a grip with the runaway code
