+++ topics/addin/Rimshot.Shared/Rimshot.Shared.projitems
@@ -23,11 +23,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\IfcGuid.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\TaskProgress.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Views\Views.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Workshop\IssueList.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Workshop\IssueListPane.xaml.cs">
-      <DependentUpon>IssueListPane.xaml</DependentUpon>
+    <Compile Include="$(MSBuildThisFileDirectory)Plugin\RimshotPlugin.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Plugin\RimshotPane.xaml.cs">
+      <DependentUpon>RimshotPane.xaml</DependentUpon>
     </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Workshop\IssueListUIBindings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Plugin\RimshotAppBindings.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="$(MSBuildThisFileDirectory)Images\icon-register.ico">
@@ -100,7 +100,7 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <Page Include="$(MSBuildThisFileDirectory)Workshop\IssueListPane.xaml">
+    <Page Include="$(MSBuildThisFileDirectory)Plugin\RimshotPane.xaml">
       <CopyToOutputDirectory>Never</CopyToOutputDirectory>
       <SubType>Designer</SubType>
       <Generator>XamlIntelliSenseFileGenerator</Generator>
